refactor(Pet Battles): Moved all logic into utils files for better clarity & development
