+++ requirements.txt
@@ -1,0 +1,7 @@
+discord
+requests
+python-dotenv
+beautifulsoup4
+pymongo
+topggpy
+aiohttp
