+++ utils/pets/petconstants.py
